refactor(devices): address React Doctor warnings
